add permissin module

+++ src/abs/middleground/technology/permission/manager/__init__.py
@@ -21,7 +21,15 @@ from abs.middleground.technology.permission.manager.register import \
 class PermissionServer(BaseManager):
 
     @classmethod
-    def create_platform(cls, name, company_id, app_type, remark, limit=4):
+    def create_platform(
+        cls,
+        name,
+        company_id,
+        company_name,
+        app_type,
+        remark,
+        limit=100
+    ):
         if PlatForm.query().filter(name=name).count() > 0:
             raise BusinessError("该平台已注册")
 
@@ -31,6 +39,7 @@ class PermissionServer(BaseManager):
         platform = PlatForm.create(
             name=name,
             company_id=company_id,
+            company_name=company_name,
             remark=remark,
             app_type=app_type,
         )
@@ -49,19 +58,32 @@ class PermissionServer(BaseManager):
         return platform_qs
 
     @classmethod
+    def search_platform(cls, current_page, **search_info):
+        platform_qs = PlatForm.query(
+            **search_info
+        ).order_by("-create_time")
+        spliter = Splitor(current_page, platform_qs)
+        return spliter
+
+    @classmethod
     def update_platform(cls, platform_id, **update_infos):
         platform = cls.get_platform(platform_id)
         platform.update(**update_infos)
         return platform
 
     @classmethod
-    def authorize(cls, platform_id, company_id, remark):
+    def remove_platform(cls, platform_id):
         platform = cls.get_platform(platform_id)
-        authorization = Authorization.create(
-            company_id=company_id,
-            platform=platform,
-            remark=remark,
-        )
+        authorization_list = Authorization.get_byplatform(platform)
+        if authorization_list:
+            raise BusinessError("该平台已有公司被授权，不可删除")
+        platform.delete()
+
+    @classmethod
+    def get_authorization(cls, authorization_id):
+        authorization = Authorization.get_byid(authorization_id)
+        if authorization is None:
+            raise BusinessError("授权不存在")
         return authorization
 
     @classmethod
@@ -69,6 +91,47 @@ class PermissionServer(BaseManager):
         authorization = Authorization.get_byappkey(appkey)
         if authorization is None:
             raise BusinessError("授权不存在")
+        return authorization
+
+    @classmethod
+    def search_authorization(cls, current_page, platform_id, **search_info):
+        platform = cls.get_platform(platform_id)
+        authorization_qs = Authorization.query(
+            platform=platform,
+            **search_info
+        ).filter(
+        )
+        spliter = Splitor(current_page, authorization_qs)
+        return spliter
+
+    @classmethod
+    def update_authorization(cls, authorization_id, **update_info):
+        authorization = cls.get_authorization(authorization_id)
+        authorization.update(
+            **update_info
+        )
+
+    @classmethod
+    def remove_authorization(cls, authorization_id):
+        authorization = cls.get_authorization(authorization_id)
+        authorization.delete()
+
+    @classmethod
+    def authorize(cls, platform_id, company_id, company_name, remark):
+        platform = cls.get_platform(platform_id)
+        is_exsited = Authorization.check_unique(
+            platform_id,
+            company_id
+        )
+        if is_exsited:
+            raise BusinessError("该公司重复授权")
+
+        authorization = Authorization.create(
+            company_id=company_id,
+            company_name=company_name,
+            platform=platform,
+            remark=remark,
+        )
         return authorization
 
     @classmethod
@@ -107,7 +170,11 @@ class PermissionServer(BaseManager):
     @classmethod
     def get_all_rule_byplatform(cls, platform_id):
         platform = cls.get_platform(platform_id)
-        return RuleHelper(platform).root.get_tree()
+        all_rule = [
+            entity.get_tree()
+            for entity in RuleHelper(platform).root.get_children()
+        ]
+        return all_rule
 
     @classmethod
     def add_rule(cls, platform_id, name, parent_id, description, remark):
@@ -129,31 +196,27 @@ class PermissionServer(BaseManager):
 
     @classmethod
     def remove_rule(cls, rule_id):
-        return cls.get_rule(rule_id).delete()
+        rule = cls.get_rule(rule_id)
+        if rule.get_children():
+            raise BusinessError("不能删除，规则下面仍有子节点!")
+        rule.delete()
 
     @classmethod
-    def get_organization(cls, organization_id):
-        organization = Organization.get_byid(organization_id)
-        if organization is None:
-            raise BusinessError("组织不存在")
-        return organization
-
-    @classmethod
-    def get_all_organization_byappkey(cls, appkey):
-        authorization = cls.get_authorization_byappkey(appkey)
-        helper = permission_register.get_helper(authorization).organization
-        all_organization = []
-        if helper.root is not None:
-            all_organization = helper.root.get_tree()
-        return all_organization
-
-    @classmethod
-    def add_organization(cls, appkey, parent_id, name, description, remark):
+    def add_organization(
+        cls,
+        appkey,
+        position_id_list,
+        parent_id,
+        name,
+        description,
+        remark
+    ):
         authorization = cls.get_authorization_byappkey(appkey)
         organization = Organization.create(
             name=name,
             parent_id=parent_id,
             description=description,
+            position_id_list=json.dumps(position_id_list),
             remark=remark,
             authorization=authorization,
         )
@@ -161,6 +224,57 @@ class PermissionServer(BaseManager):
             authorization
         ).organization.force_refresh()
         return organization
+
+    @classmethod
+    def _hung_position_byorganization(cls, organization_list):
+        position_id_set = set()
+        for organization in organization_list:
+            for position_id in json.loads(organization.position_id_list):
+                position_id_set.add(position_id)
+
+        position_map = {
+            position.id: position
+            for position in Position.search(
+                id__in=list(position_id_set)
+            )
+        }
+
+        for organization in organization_list:
+            position_list = []
+            for position_id in json.loads(organization.position_id_list):
+                position_list.append(
+                    position_map[position_id]
+                )
+            organization.position_list = position_list
+
+    @classmethod
+    def get_organization(cls, organization_id):
+        organization = Organization.get_byid(organization_id)
+        if organization is None:
+            raise BusinessError("组织不存在")
+        cls._hung_position_byorganization([organization])
+        return organization
+
+    @classmethod
+    def search_organization_byappkey(cls, appkey, current_page, **search_info):
+        authorization = cls.get_authorization_byappkey(appkey)
+        organization_qs = Organization.query(
+            authorization=authorization,
+            **search_info
+        )
+        spliter = Splitor(current_page, organization_qs)
+        cls._hung_position_byorganization(spliter.get_list())
+        return spliter
+
+    @classmethod
+    def get_all_organization_byappkey(cls, appkey):
+        authorization = cls.get_authorization_byappkey(appkey)
+        helper = permission_register.get_helper(authorization).organization
+        all_organization = [
+            entity.get_tree()
+            for entity in helper.root.get_children()
+        ]
+        return all_organization
 
     @classmethod
     def update_organization(cls, organization_id, **update_info):
@@ -191,8 +305,9 @@ class PermissionServer(BaseManager):
     def search_rule_group(cls, current_page, appkey, **search_info):
         authorization = cls.get_authorization_byappkey(appkey)
         rule_group_qs = RuleGroup.query(
-            authorization=authorization
-        ).filter(**search_info)
+            authorization=authorization,
+            **search_info
+        ).filter()
         spliter = Splitor(current_page, rule_group_qs)
         return spliter
 
@@ -226,19 +341,29 @@ class PermissionServer(BaseManager):
         return position
 
     @classmethod
+    def search_position_byappkey(cls, appkey, current_page, **search_info):
+        authorization = cls.get_authorization_byappkey(appkey)
+        position_qs = Position.query(
+            authorization=authorization,
+            **search_info
+        )
+        spliter = Splitor(current_page, position_qs)
+        return spliter
+
+    @classmethod
     def get_all_position_byappkey(cls, appkey):
         authorization = cls.get_authorization_byappkey(appkey)
         helper = permission_register.get_helper(authorization).position
-        all_position = []
-        if helper.root is not None:
-            all_position = helper.root.get_tree()
+        all_position = [
+            entity.get_tree()
+            for entity in helper.root.get_children()
+        ]
         return all_position
 
     @classmethod
     def add_position(
         cls,
         appkey,
-        organization_id,
         rule_group_id,
         parent_id,
         name,
@@ -246,14 +371,12 @@ class PermissionServer(BaseManager):
         remark
     ):
         authorization = cls.get_authorization_byappkey(appkey)
-        organization = cls.get_organization(organization_id)
         rule_group = cls.get_rule_group(rule_group_id)
         position = Position.create(
             name=name,
             description=description,
             parent_id=parent_id,
             remark=remark,
-            organization=organization,
             rule_group=rule_group,
             authorization=authorization,
         )
@@ -263,6 +386,10 @@ class PermissionServer(BaseManager):
     @classmethod
     def update_position(cls, position_id, **update_info):
         position = cls.get_position(position_id)
+        if "rule_group_id" in update_info:
+            rule_group_id = update_info.pop("rule_group_id")
+            rule_group = cls.get_rule_group(rule_group_id)
+            update_info.update({"rule_group": rule_group})
         position.update(**update_info)
         permission_register.get_helper(
             position.authorization
@@ -272,6 +399,9 @@ class PermissionServer(BaseManager):
     @classmethod
     def remove_position(cls, position_id):
         position = cls.get_position(position_id)
+        position_qs = position.get_children()
+        if position_qs.count() > 0:
+            raise BusinessError("身份存有下级节点，不能删除")
         position.delete()
         permission_register.get_helper(
             position.authorization
